+++ OBBBA/groups.sh
@@ -1,0 +1,20 @@
+#!/bin/zsh
+
+./implement.sh A
+./implement.sh B
+./implement.sh C
+./implement.sh D
+./implement.sh E
+./implement.sh F
+./implement.sh G
+./implement.sh H
+./implement.sh I
+./implement.sh J
+./implement.sh K
+./implement.sh L
+./implement.sh M
+./implement.sh N
+./implement.sh O
+./implement.sh P
+
+exit 0