+++ taxcalc/calculator.py
@@ -14,9 +14,9 @@ import paramtools
 from taxcalc.calcfunctions import (TaxInc, SchXYZTax, GainsTax, AGIsurtax,
                                    NetInvIncTax, AMT, EI_PayrollTax, Adj,
                                    DependentCare, ALD_InvInc_ec_base, CapGains,
-                                   SSBenefits, UBI, AGI, ItemDed,
-                                   StdDed, AdditionalMedicareTax, F2441, EITC,
-                                   RefundablePayrollTaxCredit,
+                                   SSBenefits, UBI, AGI, MiscDed,
+                                   ItemDed, StdDed, AdditionalMedicareTax,
+                                   F2441, EITC, RefundablePayrollTaxCredit,
                                    ChildDepTaxCredit, AdditionalCTC, CTC_new,
                                    PersonalTaxCredit, SchR,
                                    AmOppCreditParts, EducationTaxCredit,
@@ -1391,6 +1391,7 @@ class Calculator():
         CapGains(self.__policy, self.__records)
         SSBenefits(self.__policy, self.__records)
         AGI(self.__policy, self.__records)
+        MiscDed(self.__policy, self.__records)
         ItemDed(self.__policy, self.__records)
         AdditionalMedicareTax(self.__policy, self.__records)
         StdDed(self.__policy, self.__records)