+++ taxcalc/tests/test_calcfunctions.py
@@ -16,7 +16,7 @@ from taxcalc import calcfunctions
 
 class GetFuncDefs(ast.NodeVisitor):
     """
-    Return information about each function defined in the functions.py file.
+    Return information about each function defined in calcfunctions.py file.
     """
     def __init__(self):
         """
@@ -39,6 +39,8 @@ class GetFuncDefs(ast.NodeVisitor):
         """
         visit the specified FunctionDef node
         """
+        if node.name == 'SchXYZ':
+            return  # skipping SchXYZ function that has multiple returns
         self.fname = node.name
         self.fnames.append(self.fname)
         self.fargs[self.fname] = []
@@ -68,6 +70,7 @@ class GetFuncDefs(ast.NodeVisitor):
         self.generic_visit(node)
 
 
+@pytest.mark.calc_and_used_vars
 def test_calc_and_used_vars(tests_path):
     """
     Runs two kinds of tests on variables used in the calcfunctions.py file:
@@ -92,8 +95,6 @@ def test_calc_and_used_vars(tests_path):
     # .. create set of vars that are actually calculated in calcfunctions.py
     all_cvars = set()
     for fname in fnames:
-        if fname == 'BenefitSurtax':
-            continue  # because BenefitSurtax is not really a function
         all_cvars.update(set(cvars[fname]))
     # .. add to all_cvars set variables calculated in Records class
     all_cvars.update(set(['num', 'sep', 'exact']))
@@ -110,8 +111,7 @@ def test_calc_and_used_vars(tests_path):
             found_error1 = True
             msg1 += f'VAR NOT CALCULATED: {var}\n'
     # Test (2):
-    faux_functions = ['EITCamount', 'ComputeBenefit', 'BenefitPrograms',
-                      'BenefitSurtax', 'BenefitLimitation', 'SchXYZ']
+    faux_functions = ['EITCamount', 'SchXYZ', 'BenefitPrograms']
     found_error2 = False
     msg2 = 'calculated & returned variables are not function arguments\n'
     for fname in fnames:
@@ -157,8 +157,6 @@ def test_function_args_usage(tests_path):
         fname = match.group(1)
         fargs = match.group(2).split(',')  # list of function arguments
         fbody = match.group(3)
-        if fname == 'Taxes':
-            continue  # because Taxes has part of fbody in return statement
         for farg in fargs:
             arg = farg.strip()
             if fbody.find(arg) < 0:
@@ -558,6 +556,7 @@ def test_EITC(test_tuple, expected_value, skip_jit):
 
 # Parameter values for tests
 PT_qbid_rt = 0.2
+PT_qbid_limited = True
 PT_qbid_taxinc_thd = [160700.0, 321400.0, 160725.0, 160700.0, 321400.0]
 PT_qbid_taxinc_gap = [50000.0, 100000.0, 50000.0, 50000.0, 100000.0]
 PT_qbid_w2_wages_rt = 0.5
@@ -565,6 +564,8 @@ PT_qbid_alt_w2_wages_rt = 0.25
 PT_qbid_alt_property_rt = 0.025
 PT_qbid_ps = [9e99, 9e99, 9e99, 9e99, 9e99]
 PT_qbid_prt = 0.0
+PT_qbid_min_ded = 400.0  # OBBBA value in 2026
+PT_qbid_min_qbi = 1000.0  # OBBBA value in 2026
 
 # Input variable values for tests
 c00100 = [527860.66, 337675.10, 603700.00, 90700.00]
@@ -579,49 +580,58 @@ e02100 = [0.00, 0.00, 0.00, 0.00]
 e27200 = [0.00, 0.00, 0.00, 0.00]
 e00650 = [5000.00, 8000.00, 3000.00, 9000.00]
 c01000 = [7000.00, 4000.00, -3000.00, -3000.00]
+senior_deduction = [0.00, 0.00, 1000.00, 0.00]
+auto_loan_interest_deduction = [0.00, 0.00, 0.00, 1000.00]
 PT_SSTB_income = [0, 1, 1, 1]
 PT_binc_w2_wages = [0.00, 0.00, 0.00, 0.00]
 PT_ubia_property = [0.00, 0.00, 0.00, 0.00]
-c04800 = [0.0, 0.0, 0.0, 0.0]  # unimportant for function
-qbided = [0.0, 0.0, 0.0, 0.0]  # unimportant for function
+c04800 = [0.0, 0.0, 0.0, 0.0]  # calculated by function
+qbided = [0.0, 0.0, 0.0, 0.0]  # calculated by function
 
 tuple0 = (
     c00100[0], standard[0], c04470[0], c04600[0], MARS[0], e00900[0],
     c03260[0], e26270[0],
-    e02100[0], e27200[0], e00650[0], c01000[0], PT_SSTB_income[0],
-    PT_binc_w2_wages[0], PT_ubia_property[0], PT_qbid_rt,
+    e02100[0], e27200[0], e00650[0], c01000[0],
+    senior_deduction[0], auto_loan_interest_deduction[0], PT_SSTB_income[0],
+    PT_binc_w2_wages[0], PT_ubia_property[0], PT_qbid_rt, PT_qbid_limited,
     PT_qbid_taxinc_thd, PT_qbid_taxinc_gap, PT_qbid_w2_wages_rt,
-    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt, c04800[0],
-    PT_qbid_ps, PT_qbid_prt, qbided[0])
-expected0 = (490860.66, 0)
+    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt,
+    PT_qbid_ps, PT_qbid_prt, PT_qbid_min_ded, PT_qbid_min_qbi,
+    c04800[0], qbided[0])
+expected0 = (490460.66, 400.00)
 tuple1 = (
     c00100[1], standard[1], c04470[1], c04600[1], MARS[1], e00900[1],
     c03260[1], e26270[1],
-    e02100[1], e27200[1], e00650[1], c01000[1], PT_SSTB_income[1],
-    PT_binc_w2_wages[1], PT_ubia_property[1], PT_qbid_rt,
+    e02100[1], e27200[1], e00650[1], c01000[1],
+    senior_deduction[1], auto_loan_interest_deduction[1], PT_SSTB_income[1],
+    PT_binc_w2_wages[1], PT_ubia_property[1], PT_qbid_rt, PT_qbid_limited,
     PT_qbid_taxinc_thd, PT_qbid_taxinc_gap, PT_qbid_w2_wages_rt,
-    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt, c04800[1],
-    PT_qbid_ps, PT_qbid_prt, qbided[1])
+    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt,
+    PT_qbid_ps, PT_qbid_prt, PT_qbid_min_ded, PT_qbid_min_qbi,
+    c04800[1], qbided[1])
 expected1 = (284400.08, 4275.02)
 tuple2 = (
     c00100[2], standard[2], c04470[2], c04600[2], MARS[2], e00900[2],
     c03260[2], e26270[2],
-    e02100[2], e27200[2], e00650[2], c01000[2], PT_SSTB_income[2],
-    PT_binc_w2_wages[2], PT_ubia_property[2], PT_qbid_rt,
+    e02100[2], e27200[2], e00650[2], c01000[2],
+    senior_deduction[2], auto_loan_interest_deduction[2], PT_SSTB_income[2],
+    PT_binc_w2_wages[2], PT_ubia_property[2], PT_qbid_rt, PT_qbid_limited,
     PT_qbid_taxinc_thd, PT_qbid_taxinc_gap, PT_qbid_w2_wages_rt,
-    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt, c04800[2],
-    PT_qbid_ps, PT_qbid_prt, qbided[2])
-expected2 = (579300.00, 0)
+    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt,
+    PT_qbid_ps, PT_qbid_prt, PT_qbid_min_ded, PT_qbid_min_qbi,
+    c04800[2], qbided[2])
+expected2 = (577900.00, 400.00)
 tuple3 = (
     c00100[3], standard[3], c04470[3], c04600[3], MARS[3], e00900[3],
     c03260[3], e26270[3],
-
-    e02100[3], e27200[3], e00650[3], c01000[3], PT_SSTB_income[3],
-    PT_binc_w2_wages[3], PT_ubia_property[3], PT_qbid_rt,
+    e02100[3], e27200[3], e00650[3], c01000[3],
+    senior_deduction[3], auto_loan_interest_deduction[3], PT_SSTB_income[3],
+    PT_binc_w2_wages[3], PT_ubia_property[3], PT_qbid_rt, PT_qbid_limited,
     PT_qbid_taxinc_thd, PT_qbid_taxinc_gap, PT_qbid_w2_wages_rt,
-    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt, c04800[3],
-    PT_qbid_ps, PT_qbid_prt, qbided[3])
-expected3 = (57500.00, 1200)
+    PT_qbid_alt_w2_wages_rt, PT_qbid_alt_property_rt,
+    PT_qbid_ps, PT_qbid_prt, PT_qbid_min_ded, PT_qbid_min_qbi,
+    c04800[3], qbided[3])
+expected3 = (56500.00, 1200)
 
 
 @pytest.mark.parametrize(
@@ -885,4 +895,78 @@ def test_AGI(test_tuple, expected_value, skip_jit):
     """
     test_value = calcfunctions.AGI(*test_tuple)
     print('Returned from agi function: ', test_value)
+    assert np.allclose(test_value, expected_value)
+
+
+# parameters for next test
+e03150 = 0
+e03210 = 0
+c03260 = 0
+e03270 = 0
+e03300 = 0
+e03400 = 0
+e03500 = 0
+e00800 = 0
+e03220 = 0
+e03230 = 0
+e03240 = 0
+e03290 = 0
+care_deduction = 0
+ALD_StudentLoan_hc = 0
+ALD_SelfEmp_HealthIns_hc = 0
+ALD_KEOGH_SEP_hc = 0
+ALD_EarlyWithdraw_hc = 0
+ALD_AlimonyPaid_hc = 0
+ALD_AlimonyReceived_hc = 0
+ALD_EducatorExpenses_hc = 0
+ALD_HSADeduction_hc = 0
+ALD_IRAContributions_hc = 0
+ALD_DomesticProduction_hc = 0
+ALD_Tuition_hc = 0
+MARS = 1
+earned = 200000
+overtime_income = 13000
+ALD_OvertimeIncome_hc = 0.
+ALD_OvertimeIncome_c = [12500, 25000, 12500, 12500, 12500]
+ALD_OvertimeIncome_ps = [150000, 300000, 150000, 150000, 150000]
+ALD_OvertimeIncome_prt = 0.10
+tip_income = 30000
+ALD_TipIncome_hc = 0.
+ALD_TipIncome_c = [25000, 25000, 25000, 25000, 25000]
+ALD_TipIncome_ps = [150000, 300000, 150000, 150000, 150000]
+ALD_TipIncome_prt = 0.10
+c02900 = 0  # calculated in function
+ALD_OvertimeIncome = 0  # calculated in function
+ALD_TipIncome = 0  # calculated in function
+
+tuple0 = (
+    e03150, e03210, c03260,
+    e03270, e03300, e03400, e03500, e00800,
+    e03220, e03230, e03240, e03290, care_deduction,
+    ALD_StudentLoan_hc, ALD_SelfEmp_HealthIns_hc, ALD_KEOGH_SEP_hc,
+    ALD_EarlyWithdraw_hc, ALD_AlimonyPaid_hc, ALD_AlimonyReceived_hc,
+    ALD_EducatorExpenses_hc, ALD_HSADeduction_hc, ALD_IRAContributions_hc,
+    ALD_DomesticProduction_hc, ALD_Tuition_hc,
+    MARS, earned,
+    overtime_income, ALD_OvertimeIncome_hc, ALD_OvertimeIncome_c,
+    ALD_OvertimeIncome_ps, ALD_OvertimeIncome_prt,
+    tip_income, ALD_TipIncome_hc, ALD_TipIncome_c,
+    ALD_TipIncome_ps, ALD_TipIncome_prt,
+    c02900, ALD_OvertimeIncome, ALD_TipIncome
+)
+ovr = 12500 - (200000 - 150000) * 0.10
+tip = 25000 - (200000 - 150000) * 0.10
+c02900 = 0 + ovr + tip
+expected0 = (c02900, ovr, tip)
+
+
+@pytest.mark.parametrize(
+    'test_tuple,expected_value', [(tuple0, expected0)]
+)
+def test_Adj(test_tuple, expected_value, skip_jit):
+    """
+    Tests the Adj function ALD_OvertimeIncome and ALD_TipIncome code
+    """
+    test_value = calcfunctions.Adj(*test_tuple)
+    print('Returned from Adj function: ', test_value)
     assert np.allclose(test_value, expected_value)